Delete previous DataExtractor after moving all remaining functionality out of it
